En repasar guia las imagenes ya son visibles junto con el visor

+++ app/src/main/java/com/jonathanev/review/Fragments/Adaptadores/ListPreviewQuestionsAdapter.kt
@@ -9,7 +9,7 @@ import com.jonathanev.review.Fragments.ViewHolders.ListPreviewQuestionsViewHolde
 import com.jonathanev.review.databinding.ListPreviewQuestionsBinding
 
 class ListPreviewQuestionsAdapter(
-    private val clickedPlay: () -> Unit,
+    private val clickedPlay: (Int) -> Unit,
     private val clickedEdit: () -> Unit
 ): ListAdapter<PreviewQuestion, ListPreviewQuestionsViewHolder>(PreviewQuestionDiffCallback()) {
     override fun onCreateViewHolder(
